refactor: extract startup checks, address all review findings
Startup checks (EventHandlers.ts → src/startupChecks.ts):
- Extract assertStartBlocksValid + constants into dedicated module
- Gate module-level call with NODE_ENV=test skip to prevent shell env vars
  from interfering with unrelated test suites
- EventHandlers.ts re-exports all startup check symbols for test backwards compat

CI (.github/workflows/indexer-envio.yml):
- Add ENVIO_STRICT_START_BLOCK=true to codegen validation step so the guard
  throws (not warns) in CI, catching start-block misconfig before deploy

virtualPool.ts:
- Fix misleading 'handle defensively' comment on VirtualPool.Rebalanced —
  the event is real when emitted; comment now explains why handler is simpler
  than FPMM version (no oracle/RPC fetch)

AGENTS.md + indexer-envio/AGENTS.md:
- Remove blank line artifact breaking ASCII file tree

config.multichain.testnet.yaml:
- Document the mixed static+dynamic address model; note that testnet pools
  deployed before contractRegister support was added may be missing on fresh
  sync (same pattern as mainnet, same fix path: add to hardcoded list)

config.multichain.mainnet.yaml:
- Confirm OpenLiquidityStrategy is deployed on Monad mainnet (verified via
  eth_getCode — bytecode present), remove precautionary 'verify before going
  live' comment

dynamicRegistration.test.ts:
- Document why the generated module type cast is unavoidable (generated/index.d.ts
  does not export clean standalone interface types)
