+++ indexer-envio/src/EventHandlers.ts
@@ -6,77 +6,11 @@
 // modules are imported below; their registrations fire at module load time.
 // ===========================================================================
 
-// ---------------------------------------------------------------------------
-// Startup invariant: fail loudly if ENVIO_START_BLOCK is set above the
-// FPMMFactory's first deployment block for any chain. If start_block is too
-// high, FPMMDeployed events are never seen, contractRegister never fires,
-// and all pool events are silently dropped — no error, just missing data.
-//
-// First factory deployment blocks for production mainnet chains only.
-// We only validate mainnet chains here to avoid false-fatal errors when a
-// testnet env var (e.g. ENVIO_START_BLOCK_MONAD_TESTNET) is left set in
-// .env from a previous testnet run — that should never block a mainnet start.
-// Testnet configs are dev-only and not deployed to hosted Envio production.
-//
-// Celo mainnet (42220):  60668100 — initial batch of 4 FPMM pools
-// Monad mainnet (143):   60759432 — initial batch of 3 FPMM pools
-// ---------------------------------------------------------------------------
-export const FPMM_FIRST_DEPLOY_BLOCK: Record<number, number> = {
-  42220: 60668100, // Celo mainnet
-  143: 60759432, // Monad mainnet
-};
+import { runStartupChecks } from "./startupChecks";
 
-// Each mainnet chain maps to its dedicated env var.
-export const START_BLOCK_ENV_NAME: Record<number, string> = {
-  42220: "ENVIO_START_BLOCK_CELO",
-  143: "ENVIO_START_BLOCK_MONAD",
-};
-
-/**
- * Warn (or throw in CI) if an ENVIO_START_BLOCK_* env var is set above the
- * first FPMMFactory deployment block for its chain. If start_block is too
- * high, FPMMDeployed events are never seen, contractRegister never fires,
- * and pool events are silently dropped.
- *
- * Throws only when ENVIO_STRICT_START_BLOCK=true (e.g. in CI/hosted runs).
- * In local dev mode it logs a warning instead, so a leftover mainnet env var
- * doesn't hard-fail an unrelated testnet run.
- *
- * Exported for testing. Called at module load time below.
- */
-export function assertStartBlocksValid(
-  envOverrides: Record<number, string | undefined>,
-  strict = process.env.ENVIO_STRICT_START_BLOCK === "true",
-): void {
-  for (const [chainIdStr, firstDeployBlock] of Object.entries(
-    FPMM_FIRST_DEPLOY_BLOCK,
-  )) {
-    const chainId = Number(chainIdStr);
-    const envVal = envOverrides[chainId];
-    if (envVal === undefined || envVal === "") continue;
-    const startBlock = Number(envVal);
-    if (!Number.isFinite(startBlock)) continue;
-    if (startBlock > firstDeployBlock) {
-      const envVarName = START_BLOCK_ENV_NAME[chainId];
-      const msg =
-        `[EventHandlers] start block for chain ${chainId} is ${startBlock}, ` +
-        `but FPMMFactory first deployed at block ${firstDeployBlock}. ` +
-        `All factory deploy events will be missed and no pools will be indexed. ` +
-        `Lower ${envVarName} to ≤${firstDeployBlock} or remove the override.`;
-      if (strict) {
-        throw new Error(`FATAL: ${msg}`);
-      } else {
-        console.warn(`⚠️  WARNING: ${msg}`);
-      }
-    }
-  }
-}
-
-// Run the check at startup with mainnet env var values only.
-assertStartBlocksValid({
-  42220: process.env.ENVIO_START_BLOCK_CELO,
-  143: process.env.ENVIO_START_BLOCK_MONAD,
-});
+// Run startup invariant checks (skipped in NODE_ENV=test).
+// See src/startupChecks.ts for details and rationale.
+runStartupChecks();
 
 // Handler registrations (side-effect imports)
 import "./handlers/fpmm";
@@ -123,3 +57,10 @@ export {
 
 // Trading limits constant (used by decimals.test.ts)
 export { TRADING_LIMITS_INTERNAL_DECIMALS } from "./tradingLimits";
+
+// Startup checks (used by startBlockInvariant.test.ts)
+export {
+  assertStartBlocksValid,
+  FPMM_FIRST_DEPLOY_BLOCK,
+  START_BLOCK_ENV_NAME,
+} from "./startupChecks";
